+++ src/bin/commands/setup/template.js
@@ -15,10 +15,17 @@ const template = {
   "phenomic": {/* placeholder */},
   "babel": {
     "presets": [
-      "react",
-      "es2015",
-      "stage-1",
+      "babel-preset-react",
+      "babel-preset-es2015",
+      "babel-preset-stage-1",
     ],
+    "env": {
+      "production": {
+        "presets": [
+          "babel-preset-react-optimize",
+        ],
+      },
+    },
   },
   "eslintConfig": {
     "parser": "babel-eslint",